changed logic to set stop id, moved Stop to model folder
