+++ main-app/src/main/java/com/nyctransittracker/mainapp/model/StationResponse.java
@@ -1,7 +1,6 @@
 package com.nyctransittracker.mainapp.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.nyctransittracker.mainapp.entity.Stop;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
