+++ main-app/src/main/java/com/nyctransittracker/mainapp/repository/StopRepository.java
@@ -1,6 +1,6 @@
 package com.nyctransittracker.mainapp.repository;
 
-import com.nyctransittracker.mainapp.entity.Stop;
+import com.nyctransittracker.mainapp.model.Stop;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StopRepository
