+++ main-app/src/main/java/com/nyctransittracker/mainapp/service/StopService.java
@@ -1,11 +1,9 @@
 package com.nyctransittracker.mainapp.service;
 
-import com.nyctransittracker.mainapp.entity.Stop;
+import com.nyctransittracker.mainapp.model.Stop;
 import com.nyctransittracker.mainapp.repository.StopRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.Random;
 
 @Service
 @RequiredArgsConstructor
@@ -14,7 +12,7 @@ public class StopService {
     private final StopRepository repository;
     public void createStation(Stop stop, String direction, String route) {
         var newStop = new Stop();
-        newStop.setUniqueId(new Random().nextLong());
+        newStop.setUniqueId(stop.getGTFSStopID().toString() + route + direction);
         newStop.setGTFSStopID(stop.getGTFSStopID());
         newStop.setName(stop.getName());
         newStop.setLatitude(stop.getLatitude());
