+++ main-app/src/main/java/com/nyctransittracker/mainapp/util/StationFetcher.java
@@ -1,12 +1,10 @@
 package com.nyctransittracker.mainapp.util;
 
-import com.nyctransittracker.mainapp.MainAppApplication;
-import com.nyctransittracker.mainapp.entity.Stop;
+import com.nyctransittracker.mainapp.model.Stop;
 import com.nyctransittracker.mainapp.model.StationResponse;
 import com.nyctransittracker.mainapp.service.StopService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
 import org.springframework.stereotype.Component;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -20,7 +18,7 @@ public class StationFetcher implements CommandLineRunner {
     private final static String url = "https://www.goodservice.io/api/stops/";
     private final StopService service;
 
-    ObjectMapper mapper = new ObjectMapper();
+    private static ObjectMapper mapper = new ObjectMapper();
 
     @Override
     public void run(String... args) throws IOException {
